BUGFIX: Scroll to top of Page on changing pages
BUG: Whenever moving through pages, the position
of page remains the same even in the new page.
For ex, in the homepage, try opening the last article.
It should open the 404 page since the route does not exist yet.
But the 404 page is opened from the same position from where we
left the home page.

To fix this, see this Stack Overflow Answer:

https://stackoverflow.com/a/61602724/12982627